tegra-helper-scripts: update tegra210-flash-helper for R32.6.1
The L4T flash.sh script manipulates the PT partition table
before flashing, so update the helper script to do the same.

+++ recipes-bsp/tegra-binaries/tegra-helper-scripts/tegra210-flash-helper.sh
@@ -11,6 +11,36 @@ imgfile=
 dataimg=
 inst_args=""
 blocksize=4096
+bootloader_part="sdmmc_boot"
+
+generate_pt_header() {
+    local in="$1"
+    local out="$2"
+    local blpart="$3"
+    local pline blksize partnumber partname start_location partsize partfile partguid partfilltoend
+    local -a PARTS
+    mapfile PARTS < <("nvflashxmlparse" -t $blpart "$in")
+    for pline in "${PARTS[@]}"; do
+	eval "$pline"
+	if [ "$partname" = "PT" ]; then
+	    break
+	fi
+    done
+    if [ "$partname" != "PT" ]; then
+	echo "ERR: could not locate PT in flash layout" >&2
+	return 1
+    fi
+    local ptsize=$(expr $partsize \* $blksize)
+    rm -f crc-flash.xml.tmp "$out"
+    sed -e"s,$partfile,crc-$partfile," "$in" > crc-flash.xml.tmp
+    "tegraparser" --pt crc-flash.xml.tmp || return 1
+    cp crc-flash.xml.tmp "$out"
+    truncate -s $ptsize crc-$partfile
+    printf "\x01\x00\x00\x00\x00\x00\x00\x00PTHD\x00\x00\x00\x00" | \
+	dd of=crc-$partfile seek=$(expr $ptsize - 16) bs=1 count=16 conv=notrunc status=none || return 1
+    "tegrahost" --fillcrc32 crc-$partfile
+    return 0
+}
 
 ARGS=$(getopt -n $(basename "$0") -l "bup,no-flash,sdcard,spi-only,datafile:,usb-instance:" -o "u:s:b:B:yc:" -- "$@")
 if [ $? -ne 0 ]; then
@@ -92,7 +122,8 @@ imgfile="$7"
 shift 7
 
 here=$(readlink -f $(dirname "$0"))
-flashapp=$here/tegraflash.py
+PATH="$here:$PATH"
+flashapp=tegraflash.py
 
 if [ -e ./flashvars ]; then
     . ./flashvars
@@ -146,6 +177,9 @@ if [ "$boardid" = "3448" ]; then
 	boardsku="0000"
 	BOARDSKU="0000"
     fi
+    if [ "$boardsku" != "0002" ]; then
+	bootloader_part="spi"
+    fi
     for var in $FLASHVARS; do
 	eval pat=$`echo $var`
 	if [ -z "$pat" ]; then
@@ -160,6 +194,11 @@ if [ -n "$DTBFILE" ]; then
     dtb_file="$DTBFILE"
 else
     DTBFILE="$dtb_file"
+fi
+
+if [ "$spi_only" = "yes" -a "$bootloader_part" != "spi" ]; then
+    echo "ERR: --spi-only specified for eMMC platform" >&2
+    exit 1
 fi
 
 [ -f ${cvm_bin} ] && rm -f ${cvm_bin}
@@ -198,14 +237,13 @@ else
     fi
     touch APPFILE
 fi
+
+generate_pt_header "$flash_in" flash.xml.tmp "$bootloader_part" || exit 1
+
 if [ "$spi_only" = "yes" ]; then
-    if [ ! -e "$here/nvflashxmlparse" ]; then
-	echo "ERR: missing nvflashxmlparse script" >&2
-	exit 1
-    fi
-    "$here/nvflashxmlparse" --extract -t spi -o flash.xml.tmp "$flash_in" || exit 1
-else
-    cp "$flash_in" flash.xml.tmp
+    cp flash.xml.tmp flash.xml.tmp-1
+    "nvflashxmlparse" --extract -t spi -o flash.xml.tmp flash.xml.tmp-1 || exit 1
+    rm flash.xml.tmp-1
 fi
 sed -e"s,VERFILE,${MACHINE}_bootblob_ver.txt," -e"s,DTBFILE,$DTBFILE," $appfile_sed flash.xml.tmp > flash.xml
 rm flash.xml.tmp
@@ -217,10 +255,10 @@ if [ $bup_blob -ne 0 -o "$sdcard" = "yes" ]; then
 else
     if [ -z "$sdcard" -a $no_flash -eq 0 ]; then
 	rm -f "$appfile"
-	$here/mksparse -b ${blocksize} --fillpattern=0 "$imgfile"  "$appfile" || exit 1
+	mksparse -b ${blocksize} --fillpattern=0 "$imgfile"  "$appfile" || exit 1
 	if [ -n "$datafile" ]; then
 	    rm -f "$datafile"
-	    $here/mksparse -b ${blocksize} --fillpattern=0 "$dataimg" "$datafile" || exit 1
+	    mksparse -b ${blocksize} --fillpattern=0 "$dataimg" "$datafile" || exit 1
 	fi
     fi
     cmd=${flash_cmd:-"flash;reboot"}
@@ -254,7 +292,7 @@ if [ -n "$keyfile" ]; then
     if [ $no_flash -ne 0 ]; then
 	rm -f flashcmd.txt
 	echo "#!/bin/sh" > flashcmd.txt
-	echo "python3 $flashapp ${inst_args} --bl cboot.bin.signed --bct \"$(basename $sdramcfg_file .cfg).bct\" --odmdata $odmdata \
+	echo "python3 $(basename $flashapp) ${inst_args} --bl cboot.bin.signed --bct \"$(basename $sdramcfg_file .cfg).bct\" --odmdata $odmdata \
 --bldtb \"${dtb_file}.signed\" --applet rcm_1_signed.rcm --cfg flash.xml --chip 0x21 \
 --cmd \"secureflash;reboot\" $binargs" > flashcmd.txt
 	chmod +x flashcmd.txt
